+++ docs/CHANGELOG.md
@@ -9,6 +9,7 @@ Changelog
 1.8.15 (Unreleased)
 -------------------
 - Enh: Automated code refactoring for HumHub 1.18.0-beta.6 using Rector
+- Fix #689: Fix calendar defaults and recurring visibility
 
 1.8.14 (May 21, 2026)
 ---------------------